+++ src/main/java/yoyo/Parser.java
@@ -1,3 +1,5 @@
+package yoyo;
+
 public class Parser {
     public static class Parsed {
         public final String cmd;
