+++ src/main/java/yoyo/YoyoException.java
@@ -1,3 +1,5 @@
+package yoyo;
+
 public class YoyoException extends Exception {
     public YoyoException(String msg) { super(msg); }
 }
